backend/requirements.txt
google-adk>=1.19.0
google-generativeai>=0.8.5
requests>=2.31.0
geopy>=2.4.0
wikipedia-api>=0.6.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
